pyproject.toml
[build-system]
requires = [
    'hatchling',
    'hatch-regex-commit',
]
build-backend = 'hatchling.build'


[project]
name = 'AnsibleEnvVars'
authors = [
    {name = 'Randy Syring', email = 'randy.syring@level12.io'},
]
requires-python = '>=3.12'
dynamic = ['version']
readme = 'readme.md'
dependencies = [
    "ansible-core>=2.18.6",
]


[dependency-groups]
# Note: keeping Coppy deps grouped separate from app deps should help avoid unnecessary
# conflicts when upgrading to the latest Coppy template.
dev = [
    # From Coppy:
    {include-group = "tests"},
    {include-group = "pre-commit"},
    {include-group = "audit"},
    {include-group = "nox"},
    'click>=8.1.8',
    'hatch>=1.14.0',
    'ruff>=0.9.6',

    # App specific:
    # TODO: fill in app deps here
]
# Used by nox
tests = [
    # From Coppy:
    'pytest>=8.3.4',
    'pytest-cov>=6.0',

    # App specific:
    # TODO: fill in app deps here
]
# Used by nox
pre-commit = [
    # From Coppy:
    'pre-commit>=4.1.0',
    'pre-commit-uv>=4.1.4',
]
# Used by nox
audit = [
    # From Coppy:
    'pip-audit>=2.8.0',
]
# Used by CI
nox = [
    'nox>=2025.2.9',
]
